appletlib.c: make it actally follow _BB_SUID_ALWAYS rules adduser: implement -S and code shrink / fix uid selection *: sanitize getspnam_r use
   text    data     bss     dec     hex filename
 777042     974    9676  787692   c04ec busybox_old
 776883     974    9676  787533   c044d busybox_unstripped

+++ loginutils/addgroup.c
@@ -56,7 +56,7 @@ static void new_group(char *group, gid_t gid)
 	/* add entry to group */
 	file = xfopen(bb_path_group_file, "a");
 	/* group:passwd:gid:userlist */
-	fprintf(file, "%s:x:%d:\n", group, gr.gr_gid);
+	fprintf(file, "%s:x:%u:\n", group, (unsigned)gr.gr_gid);
 	if (ENABLE_FEATURE_CLEAN_UP)
 		fclose(file);
 #if ENABLE_FEATURE_SHADOWPASSWDS
